Fix excessive library dependencies (continued)
Make sure we only link with kerberos llibraries if kerberos features
are enabled. Otherwise we produce a binaruy that requires unused
libraries, which is a liability when we try to use it on a system
where they are not installed.

Previous changed failed to update configure. Here is that fixed.